move callbacks registering to another file for recommendations
