+++ main.c
@@ -137,25 +137,6 @@ static void on_window_main_destroy(GtkWidget *widget, App *app) {
     gtk_main_quit();
 }
 
-void on_btn_try_something_new_next_clicked(GtkButton *button, App *app) {
-    app->current_recommendation_index++;
-    recommendation_service_display_recipe_try_something_new(app->current_recommendation_index, app);
-}
-
-void on_btn_try_something_new_prepare_clicked(GtkButton *button, App *app) {
-
-}
-
-void on_btn_get_recommendations_next_clicked(GtkButton *button, App *app) {
-    app->current_recommendation_index++;
-    recommendation_service_display_recipe_get_recommendations(app->current_recommendation_index, app);
-
-}
-
-static void on_btn_get_recommendations_prepare_clicked(GtkButton *button, App *app) {
-
-}
-
 static void on_btn_main_stack_recipes_clicked(GtkButton *button, App *app) {
 
 }
@@ -166,18 +147,6 @@ static void on_btn_main_stack_ingredients_clicked(GtkButton *button, App *app) {
 
 static void on_btn_main_stack_recommendations_clicked(GtkButton *button, App *app) {
     gtk_stack_set_visible_child_name(app->stack_main, "recommendation_stack");
-}
-
-static void on_btn_stack_recommendations_get_recommendations_clicked(GtkButton *button, App *app) {
-    app->current_recommendation_index = 0;
-    recommendation_service_display_recipe_get_recommendations(app->current_recommendation_index, app);
-    gtk_stack_set_visible_child_name(app->stack_recommendations, "get_recommendations");
-}
-
-static void on_btn_stack_recommendations_try_something_new_clicked(GtkButton *button, App *app) {
-    app->current_recommendation_index = 0;
-    recommendation_service_display_recipe_try_something_new(app->current_recommendation_index, app);
-    gtk_stack_set_visible_child_name(app->stack_recommendations, "try_something_new");
 }
 
 static void on_btn_modify_ingredient_add_clicked(GtkButton *button, App *app) {
@@ -253,24 +222,12 @@ static void main_gtk(int argc, char *argv[]) {
     app->current_recommendation_index = 0;
 
     gtk_builder_add_callback_symbol(builder, "on_window_main_destroy", G_CALLBACK(on_window_main_destroy));
-    gtk_builder_add_callback_symbol(builder, "on_btn_try_something_new_next_clicked",
-                                    G_CALLBACK(on_btn_try_something_new_next_clicked));
-    gtk_builder_add_callback_symbol(builder, "on_btn_try_something_new_prepare_clicked",
-                                    G_CALLBACK(on_btn_try_something_new_prepare_clicked));
-    gtk_builder_add_callback_symbol(builder, "on_btn_get_recommendations_next_clicked",
-                                    G_CALLBACK(on_btn_get_recommendations_next_clicked));
-    gtk_builder_add_callback_symbol(builder, "on_btn_get_recommendations_prepare_clicked",
-                                    G_CALLBACK(on_btn_get_recommendations_prepare_clicked));
     gtk_builder_add_callback_symbol(builder, "on_btn_main_stack_recipes_clicked",
                                     G_CALLBACK(on_btn_main_stack_recipes_clicked));
     gtk_builder_add_callback_symbol(builder, "on_btn_main_stack_ingredients_clicked",
                                     G_CALLBACK(on_btn_main_stack_ingredients_clicked));
     gtk_builder_add_callback_symbol(builder, "on_btn_main_stack_recommendations_clicked",
                                     G_CALLBACK(on_btn_main_stack_recommendations_clicked));
-    gtk_builder_add_callback_symbol(builder, "on_btn_stack_recommendations_get_recommendations_clicked",
-                                    G_CALLBACK(on_btn_stack_recommendations_get_recommendations_clicked));
-    gtk_builder_add_callback_symbol(builder, "on_btn_stack_recommendations_try_something_new_clicked",
-                                    G_CALLBACK(on_btn_stack_recommendations_try_something_new_clicked));
     gtk_builder_add_callback_symbol(builder, "on_btn_modify_ingredient_add_clicked",
                                     G_CALLBACK(on_btn_modify_ingredient_add_clicked));
     gtk_builder_add_callback_symbol(builder, "on_btn_modify_ingredient_cancel_clicked",
@@ -285,6 +242,9 @@ static void main_gtk(int argc, char *argv[]) {
                                     G_CALLBACK(on_btn_ingredients_list_modify_clicked));
     gtk_builder_add_callback_symbol(builder, "on_btn_ingredients_list_add_clicked",
                                     G_CALLBACK(on_btn_ingredients_list_add_clicked));
+
+    recommendation_service_register_callbacks(builder);
+
     gtk_builder_connect_signals(builder, app);
 
     g_object_unref(builder);
