+++ recommendation_service_controller.h
@@ -7,6 +7,7 @@
 
 #include <recommendation_service.h>
 #include <recipe_service.h>
+#include <gtk/gtk.h>
 #include "app.h"
 
 void recommendation_service_display_main_menu();
@@ -16,5 +17,7 @@ void recommendation_service_handle_option(char option, RecommendationService ser
 void recommendation_service_display_recipe_get_recommendations(int index, App *app);
 
 void recommendation_service_display_recipe_try_something_new(int index, App *app);
+
+void recommendation_service_register_callbacks(GtkBuilder *builder);
 
 #endif //CFOOD_RECOMMENDATION_SERVICE_CONTROLLER_H
